new endpoint for extended bid info of user bids

+++ Contracts/Responses/bid/BasicBidResponse.cs
@@ -6,7 +6,7 @@
 
         public DateTime Time { get; init; }
 
-        public float Amount { get; init; }
+        public decimal Amount { get; init; }
 
         public long ItemId { get; init; }
